test(models): 迁移模型目录测试
